Revert "update to use createTableResolver"
This reverts commit c699150f395cd9b0b001e70e4c04ed5eb5e4de80.
